+++ source_gen/source_gen/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `LibraryReader.pathToUrl(Uri|String)`, which computes the `import` or
   `export` path necessary to reach the provided URL from the current library.
+  Also added `pathToAsset` and `pathToElement` as convenience functions.
 
 ## 0.7.3